Feat [#15] controller, dto 형태 일부 엔티티 변화에 따른 변경 및 타입에 맞게 구현  + #13 업데이트로 인한 conflict 해결을 위한 작업 진행
- user의 경우 일단 주석으로 처리하고 뒤에서 mockUserId로 처리해서 진행

+++ morib/src/main/java/org/morib/server/api/timerView/dto/TodoCardResponseDto.java
@@ -2,5 +2,5 @@ package org.morib.server.api.timerView.dto;
 
 import java.util.List;
 
-public record TodoCardResponseDto(int totalTimeOfToday,List<TaskInTodoCardDto> task) {
+public record TodoCardResponseDto(Long totalTimeOfToday,List<TaskInTodoCardDto> task) {
 }
